address PR review: isolate unsafe hibernate, drop VMGS dance, guard Hyper-V
- guest_test_uefi: move the unsafe hibernate() into its own module, leaving the
  action selector safe.
- vmm_tests: drop the unnecessary get_guest_state_file copy dance in the token
  test (the persistent VMGS is used in place on OpenVMM).
- petri: the Hyper-V backend now errors if hibernation is requested rather than
  silently ignoring it.
