+++ guest_test_uefi/src/uefi/action.rs
@@ -9,9 +9,6 @@
 //! to run the normal test suite; specific values select alternate behaviors
 //! such as requesting hibernation.
 
-// UNSAFETY: Raw port I/O / SMC needed to request an ACPI/PSCI power transition.
-#![expect(unsafe_code)]
-
 use uefi::cstr16;
 use uefi::runtime;
 use uefi::runtime::VariableVendor;
@@ -45,46 +42,5 @@ pub fn selected_action() -> GuestAction {
     match runtime::get_variable_boxed(ACTION_VAR_NAME, &VariableVendor::GLOBAL_VARIABLE) {
         Ok((value, _)) => GuestAction::from_value(&value),
         Err(_) => GuestAction::RunTests,
-    }
-}
-
-/// Request hibernation from the platform and do not return.
-///
-/// On x86_64 this writes the ACPI PM control register requesting S4; on aarch64
-/// it issues the PSCI `SYSTEM_OFF2` call requesting hibernation. Under OpenHCL
-/// these are trapped by the paravisor, which records the hibernate token and
-/// notifies the host.
-pub fn hibernate() -> ! {
-    uefi::println!("guest_test_uefi: requesting hibernation");
-
-    #[cfg(target_arch = "x86_64")]
-    // SAFETY: Writing the emulated ACPI PM control register (PM base 0x400 +
-    // control offset 0x04) with SLP_EN set and suspend type 1 requests S4
-    // (hibernate). The write has no memory effects on the guest.
-    unsafe {
-        core::arch::asm!(
-            "out dx, ax",
-            in("dx") 0x404u16,
-            in("ax") 0x2400u16,
-            options(nomem, nostack, preserves_flags),
-        );
-    }
-
-    #[cfg(target_arch = "aarch64")]
-    // SAFETY: PSCI SYSTEM_OFF2 (0x8400_0015) with HIBERNATE_OFF (1) requests
-    // hibernation and does not return.
-    unsafe {
-        core::arch::asm!(
-            "smc #0",
-            in("x0") 0x8400_0015u64,
-            in("x1") 1u64,
-            options(nomem, nostack, preserves_flags),
-        );
-    }
-
-    // The platform should have halted this VP; if control returns, spin so we
-    // never fall through to unrelated code.
-    loop {
-        core::hint::spin_loop();
     }
 }
